refactor(cli): extract MergedConfig and config helpers into config_resolver.rs
Move MergedConfig struct, load_config, merge_string_lists, and merge_ignore_cves
from main.rs into src/cli/config_resolver.rs. Migrate corresponding tests.
All items are made pub for use by merge_config in main.rs.

Closes #354
